+++ src/OctoshiftCLI.IntegrationTests/GithubToGithub.cs
@@ -31,8 +31,7 @@ namespace OctoshiftCLI.IntegrationTests
             _helper = new TestHelper(_output, _githubApi, _githubClient);
         }
 
-        // Tracking Issue: https://github.com/github/octoshift/issues/3606
-        [Fact(Skip = "random 404 errors")]
+        [Fact]
         public async Task Basic()
         {
             var githubSourceOrg = $"e2e-testing-source-{_helper.GetOsName()}";
@@ -47,28 +46,6 @@ namespace OctoshiftCLI.IntegrationTests
             await _helper.CreateGithubRepo(githubSourceOrg, repo2);
 
             await _helper.RunGeiCliMigration($"generate-script --github-source-org {githubSourceOrg} --github-target-org {githubTargetOrg}");
-
-            await _helper.AssertGithubRepoExists(githubTargetOrg, repo1);
-            await _helper.AssertGithubRepoExists(githubTargetOrg, repo2);
-            await _helper.AssertGithubRepoInitialized(githubTargetOrg, repo1);
-            await _helper.AssertGithubRepoInitialized(githubTargetOrg, repo2);
-        }
-
-        [Fact]
-        public async Task BasicWithSsh()
-        {
-            var githubSourceOrg = $"e2e-testing-source-{_helper.GetOsName()}";
-            var githubTargetOrg = $"e2e-testing-{_helper.GetOsName()}";
-            var repo1 = "repo-1";
-            var repo2 = "repo-2";
-
-            await _helper.ResetGithubTestEnvironment(githubSourceOrg);
-            await _helper.ResetGithubTestEnvironment(githubTargetOrg);
-
-            await _helper.CreateGithubRepo(githubSourceOrg, repo1);
-            await _helper.CreateGithubRepo(githubSourceOrg, repo2);
-
-            await _helper.RunGeiCliMigration($"generate-script --github-source-org {githubSourceOrg} --github-target-org {githubTargetOrg} --ssh");
 
             await _helper.AssertGithubRepoExists(githubTargetOrg, repo1);
             await _helper.AssertGithubRepoExists(githubTargetOrg, repo2);