+++ src/ado2gh/EnvironmentVariableProvider.cs
@@ -1,4 +1,7 @@
 using System;
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleToAttribute("OctoshiftCLI.Tests")]
 
 namespace OctoshiftCLI.AdoToGithub;
 
@@ -9,19 +12,34 @@ public class EnvironmentVariableProvider
 
     private readonly OctoLogger _logger;
 
-    public EnvironmentVariableProvider(OctoLogger logger)
+    private readonly Func<string, string> _getEnvironmentVariable;
+
+    public EnvironmentVariableProvider(OctoLogger logger) : this(logger, v => Environment.GetEnvironmentVariable(v))
     {
-        _logger = logger;
     }
 
-    public virtual string GithubPersonalAccessToken() => GetSecret(GH_PAT);
+    internal EnvironmentVariableProvider(OctoLogger logger, Func<string, string> getEnvironmentVariable)
+    {
+        _logger = logger;
+        _getEnvironmentVariable = getEnvironmentVariable;
+    }
 
-    public virtual string AdoPersonalAccessToken() => GetSecret(ADO_PAT);
+    public virtual string GithubPersonalAccessToken() =>
+            GetSecret(GH_PAT)
+            ?? throw new OctoshiftCliException($"{GH_PAT} environment variable is not set.");
+
+    public virtual string AdoPersonalAccessToken() =>
+            GetSecret(ADO_PAT)
+            ?? throw new OctoshiftCliException($"{ADO_PAT} environment variable is not set.");
 
     private string GetSecret(string secretName)
     {
-        var secret = Environment.GetEnvironmentVariable(secretName) ??
-                     throw new OctoshiftCliException($"{secretName} environment variable is not set.");
+        var secret = _getEnvironmentVariable(secretName);
+
+        if (string.IsNullOrEmpty(secret))
+        {
+            return null;
+        }
 
         _logger?.RegisterSecret(secret);