+++ src/gei/Commands/GenerateScriptCommand.cs
@@ -29,12 +29,12 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
             var githubSourceOrgOption = new Option<string>("--github-source-org")
             {
                 IsRequired = false,
-                Description = "Uses GH_SOURCE_PAT env variable. Will fall back to GH_PAT if not set."
+                Description = "Uses GH_SOURCE_PAT env variable or --github-source-pat option. Will fall back to GH_PAT if not set."
             };
             var adoSourceOrgOption = new Option<string>("--ado-source-org")
             {
                 IsRequired = false,
-                Description = "Uses ADO_PAT env variable."
+                Description = "Uses ADO_PAT env variable or --ado-pat option."
             };
             var adoTeamProject = new Option<string>("--ado-team-project")
             {
@@ -42,8 +42,7 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
             };
             var githubTargetOrgOption = new Option<string>("--github-target-org")
             {
-                IsRequired = true,
-                Description = "Uses GH_PAT env variable."
+                IsRequired = true
             };
 
             // GHES migration path
@@ -77,6 +76,14 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
                 IsRequired = false,
                 Description = "Waits for each migration to finish before moving on to the next one."
             };
+            var githubSourcePath = new Option<string>("--github-source-pat")
+            {
+                IsRequired = false
+            };
+            var adoPat = new Option<string>("--ado-pat")
+            {
+                IsRequired = false
+            };
             var verbose = new Option("--verbose")
             {
                 IsRequired = false
@@ -94,9 +101,11 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
             AddOption(outputOption);
             AddOption(ssh);
             AddOption(sequential);
+            AddOption(githubSourcePath);
+            AddOption(adoPat);
             AddOption(verbose);
 
-            Handler = CommandHandler.Create<string, string, string, string, FileInfo, string, string, bool, bool, bool, bool>(Invoke);
+            Handler = CommandHandler.Create<string, string, string, string, FileInfo, string, string, bool, bool, bool, string, string, bool>(Invoke);
         }
 
         public async Task Invoke(
@@ -110,6 +119,8 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
           bool noSslVerify = false,
           bool ssh = false,
           bool sequential = false,
+          string githubSourcePat = null,
+          string adoPat = null,
           bool verbose = false)
         {
             _log.Verbose = verbose;
@@ -161,6 +172,14 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
             {
                 _log.LogInformation("SEQUENTIAL: true");
             }
+            if (githubSourcePat is not null)
+            {
+                _log.LogInformation("GITHUB SOURCE PAT: ***");
+            }
+            if (adoPat is not null)
+            {
+                _log.LogInformation("ADO PAT: ***");
+            }
 
             if (string.IsNullOrWhiteSpace(githubSourceOrg) && string.IsNullOrWhiteSpace(adoSourceOrg))
             {
@@ -168,8 +187,8 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
             }
 
             var script = string.IsNullOrWhiteSpace(githubSourceOrg) ?
-                await InvokeAdo(adoSourceOrg, adoTeamProject, githubTargetOrg, sequential) :
-                await InvokeGithub(githubSourceOrg, githubTargetOrg, ghesApiUrl, azureStorageConnectionString, noSslVerify, sequential);
+                await InvokeAdo(adoSourceOrg, adoTeamProject, githubTargetOrg, sequential, adoPat) :
+                await InvokeGithub(githubSourceOrg, githubTargetOrg, ghesApiUrl, azureStorageConnectionString, noSslVerify, sequential, githubSourcePat);
 
             if (output != null)
             {
@@ -177,18 +196,17 @@ namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
             }
         }
 
-        private async Task<string> InvokeGithub(string githubSourceOrg, string githubTargetOrg, string ghesApiUrl, string azureStorageConnectionString, bool noSslVerify, bool sequential)
+        private async Task<string> InvokeGithub(string githubSourceOrg, string githubTargetOrg, string ghesApiUrl, string azureStorageConnectionString, bool noSslVerify, bool sequential, string githubSourcePat)
         {
-            var targetApiUrl = "https://api.github.com";
-            var repos = await GetGithubRepos(_sourceGithubApiFactory.Create(targetApiUrl), githubSourceOrg);
+            var repos = await GetGithubRepos(_sourceGithubApiFactory.Create(sourcePersonalAccessToken: githubSourcePat), githubSourceOrg);
             return sequential
                 ? GenerateSequentialGithubScript(repos, githubSourceOrg, githubTargetOrg, ghesApiUrl, azureStorageConnectionString, noSslVerify)
                 : GenerateParallelGithubScript(repos, githubSourceOrg, githubTargetOrg, ghesApiUrl, azureStorageConnectionString, noSslVerify);
         }
 
-        private async Task<string> InvokeAdo(string adoSourceOrg, string adoTeamProject, string githubTargetOrg, bool sequential)
+        private async Task<string> InvokeAdo(string adoSourceOrg, string adoTeamProject, string githubTargetOrg, bool sequential, string adoPat)
         {
-            var repos = await GetAdoRepos(_sourceAdoApiFactory.Create(), adoSourceOrg, adoTeamProject);
+            var repos = await GetAdoRepos(_sourceAdoApiFactory.Create(adoPat), adoSourceOrg, adoTeamProject);
             return sequential
                 ? GenerateSequentialAdoScript(repos, adoSourceOrg, githubTargetOrg)
                 : GenerateParallelAdoScript(repos, adoSourceOrg, githubTargetOrg);